+++ examples/restful-user-service.go
@@ -8,7 +8,7 @@ import (
 )
 
 // This example is functionally the same as the example in restful-user-resource.go
-// with the only difference that is served using its the restful.DefaultContainer
+// with the only difference that is served using the restful.DefaultContainer
 
 type User struct {
 	Id, Name string